backend/pyproject.toml
[project]
name = "llm-gamebook"
version = "0.0.0"
description = "An interactive storytelling framework using LLMs to generate dynamic narratives"
license = "AGPL-3.0-or-later"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "aiosqlite>=0.22.1",
  "casefy>=1.1.0",
  "fastapi[standard]>=0.128.1",
  "greenlet>=3.3.1",
  "jinja2>=3.1.6",
  "pydantic>=2.12.5",
  "pydantic-ai-slim[anthropic,openai,google,mistral,xai]>=1.54.0",
  "platformdirs>=4.5.1",
  "pyparsing>=3.3.2",
  "pyyaml>=6.0.3",
  "sqlmodel>=0.0.32",
  "textual>=7.5.0",
  "typer>=0.21.1",
]

[project.urls]
source = "https://github.com/buzz/llm-gamebook"

[dependency-groups]
dev = [
  "mypy>=1.19.1",
  "pytest>=9.0.2",
  "pytest-asyncio>=1.3.0",
  "python-semantic-release>=10.5.3",
  "ruff>=0.15.0",
  "types-pyyaml>=6.0.12.20250915",
]

[tool.ruff]
line-length = 100
output-format = "concise"
preview = true
exclude = [".venv/**"]

[tool.ruff.lint]
# https://docs.astral.sh/ruff/rules/
select = [
  "F",  # pyflakes
  "E",  # pycodestyle errors
  "W",  # pycodestyle warnings
  "PL", # pylint

  # flake8
  "A",     #   -builtins
  "ASYNC", #   -async
  "B",     #   -bugbear
  "BLE",   #   -blind-except
  "C4",    #   -comprehensions
  "COM",   #   -commas
  "DTZ",   #   -datetimez
  "EM",    #   -errmsg
  "EXE",   #   -executable
  "FA",    #   -future-annotations
  "FBT",   #   -boolean-trap
  "G",     #   -logging-format
  "ICN",   #   -import-conventions
  "INP",   #   -no-pep420
  "INT",   #   -gettext
  "ISC",   #   -implicit-str-concat
  "LOG",   #   -logging
  "PIE",   #   -pie
  "PT",    #   -pytest-style
  "PYI",   #   -pyi
  "Q",     #   --quotes
  "RET",   #   -return
  "RSE",   #   -raise
  "SIM",   #   -simplify
  "SLOT",  #   -slots
  "T201",  #   -print
  "T10",   #   -debugger
  "TCH",   #   -type-checking
  "TD004", #   -todos
  "TD005",
  "TD006",
  "TD007",
  "TID",   #   -tidy-imports
  "YTT",   #   -2020

  "FLY",  # flynt
  "FURB", # refurb
  "I",    # isort
  "N",    # pep8-naming
  "PERF", # Perflint
  "PGH",  # pygrep-hooks
  "RUF",  # Ruff-specific rules
  "TRY",  # tryceratops
  "UP",   # pyupgrade
]
ignore = ["B010", "COM812", "PLR1702", "PLR2004", "PLR6301", "RUF029"]

[tool.ruff.lint.pylint]
max-args = 6
max-returns = 8

[tool.mypy]
plugins = ["pydantic.mypy"]
strict = true
disallow_any_explicit = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_test_loop_scope = "module"
asyncio_default_fixture_loop_scope = "module"

[tool.semantic_release]
build_command = """
    uv lock --upgrade-package "$PACKAGE_NAME"
    git add uv.lock
    uv build
"""
commit_author = "buzz <buzz@users.noreply.github.com>"
commit_message = "Version {version}"
tag_format = "v{version}"

[tool.semantic_release.changelog.default_templates]
changelog_file = "../CHANGELOG.md"

[tool.semantic_release.branches.main]
match = "main"

[tool.semantic_release.remote]
type = "github"
ignore_token_for_push = true
